Calibjo/requirements.txt
Flask==2.3.3
gunicorn==23.0.0
google-genai
python-dotenv
flask-cors